pyproject.toml
[tool.poetry]
name = "icodes"
version = "0.1.0"
description = "Intelligent Commit Ontology Distiller and Enhanced Search"
authors = ["Jordan Dimov <jdimov@a115.co.uk>", "Bayo Ade <bayo.300@gmail.com>"]
license = "CC0-1.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.12.0"
gitpython = "^3.1.43"
loguru = "^0.7.2"
openai = "^1.14.3"
pydantic-settings = "^2.2.1"
tenacity = "^8.2.3"
tiktoken = "^0.6.0"
sqlmodel = "^0.0.16"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ruff = "^0.3.5"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
